f/mmcdaniel-multiPlot: Create scrollable area that responds to checkbox
Added a scrollable test area that uses blood chemistry data requests for
testing.  The area uses a GridView with the cell size defined to
generate a 4x4 viewable area.  When blood chemistry data requests are
toggled on/off, the name of the request is added to the scrollable area.
The next step is to replace names with plot areas.