f/mmcdaniel-multiPlot: Update new point handling and plot resizing
Moved new point handling functionality from GraphArea to UIPlotSeries on
individual chart basis.  Points that are no longer in the window view
due to domain re-scaling are removed from series to cut down on storage.

Added signal handling for window resizing so that plots resize as well
and continue to fill grid.