f/mmcdaniel-actionMenu: Correct point removal in UIPlotSeries
Noticed during trial runs that plots initialized after time 0 were
having points deleted before the viewable range was filled.  This
occured because the tickCount was being used to determine when to move
points, instead of the length of the series itself.  Using the series
count to determine when to remove points resolved the issue.