f/mmcdaniel-actionMenu: Transition plotter to use createSeries function
If we do not instantiate a line series a priori, but create it when the
UIPlotSeries object is initialized, then we can easily extend this to
creating multiple line series on the same chart (e.g. for blood
pressure)

Changed extreme hemorrhage to high bleed rate so that it's easy to see
plot changes over short test runs.