f/mmcdaniel-actionMenu:  Add multi-selection support to action drawer
Actions can be highlighted by clicking or pressing enter.  Active
actions that are selected again are deactivated.  Multiple actions can
be seleted.

Successfully bound hemorrhage action to a button on the action menu.
The current plan is to bind each action's Q_Invokable function to its
corresponding entry on the List Model so that buttons that are created
and be directly connected to those functions.