f/mmcdaniel-actionEditor: Add maxWidth property to editor items
This property gives finer control over the size of an object in a grid
layout.  Thus far, I have found the best thing to do for placing
numerous items in a grid is to set fillWidth to true and place a cap on
the maximum width.  Qml documentation says that if fillWidth is true,
the item will be as wide as possible while respecting constraints (i.e.
the max width).  With this convention, we can make sure that we fill
available space in grid while still respeting boundaries.