mmcdaniel-subPlots : Create SubstanceData and Substance types
The Substance struct holds information about current state of substance
in body (blood concentration, clearance, etc.)

The SubstanceData class implements a QMap that uses substance name as
key and Substance struct as mapped item.  This map is passed as a
QVariantMap in substanceDataChanged signal to QML.  Documentaion states
that QML automatically converts QVariantMap to Javascript object.
Successfully caught signal and printed off substance names and
concentrations with this method.