mmcdaniel-subPlots: Improve sub menu and activate sub plotting
--Changed substance menu model to list model (as opposed to string
array).  This change was made because qml has built-in signal handling
for addition of a new item to ListModel (which Instantiator needs to
create new objects), whereas simple string array does not.
--Added updateSubstanceSeries function to UIPlotSeries.  Needs to be
distint from patientMetrics update because we have to search for
substance name key followed by a property key.