f/mmcdaniel-patientWizard: Re-load patient state menu on new state added
Clear patient list model and re-scan directory when user saves a states.
This keeps the patient state menu up-to-date and ordered.

Wait to set model in menu repeater until after instantiator has signaled
"objectAdded".  Doing so prevents repeater from trying to read a
non-existent model when a patient is removed from the list model linked
to the instantiator.