f/mmcdaniel-compoundWizard: Force grid view layout in wizard edit mode
Patient and nutrition editors were not loading all data when editing an
existing file.  They would load "Name" and nothing else.  Debugging
showed that not all delegate instances of grid view were being completed
before the call to load data, meaning that they lacked the requisite
signal connections to process the data. The forceLayout function forces GridView
to update all components, so calling this function makes sure that all
connections are established before loading.

It is unclear to me what caused this to become an issue (not a problem
on commit 772199fc435