f/mmcdaniel-subWizard: Fix scalar text field and verify sub data
Data tracking of scalar field entries was not behaving as intended when
writing data to field and then deleting it. The Validator on the field
suppressed the editingFinished signal and also prevented update of bound
input variable.  The result was that object tracking data would maintain
old value even after manually deleting text (this issue was not present
when using dialog "Reset" button).

Set up functions to verify permissable configuration in each tab of sub wizard.