f/mmcdaniel-subWizard: Condense and correct Package delegate definition
Previous implementation unnecessarily created extra components each time
delegate was instantiated.  This caused the signal binding to be
incorrect for delegates that were not on the first tab of the editor.Problem fixed by standing up a single UnitScalarEntry component and parenting it to the appropriate model view using state parent change.

To Do : Experimenting with adjusting visibility of clearance delegates
depending on type (systemic vs renal dynamics).  It works, but space is
still reserved for the invisible delegates (not optimal).