f/mmcdaniel-subWizard: Resolve reset of zwitterion substances
Resetting of secondary PKA field needed additional logic added for
special cases where a loaded substance had a second pka that the user
then cleared. Formerly, re-creating and setting the secondary pka field
was resulting in a type error.  This would because of impproper attempt
to remove items from "persisted items" group.

Resolved by adding secondary PKA to initial list model (so that it is in
persistent items group from the start), but pulling from
pharmacokinetics group.  Toggling zwitterion state off/on moves the
secondary pka out of and back into pk group, which effects visibility
without destroying the item (because it is still held in persistent
items group).