f/mmcdaniel-scenarioLoad: Resolve event duration error
Continuous actions that are started and never deactivated during a
scenario (like an infusion) were not being assigned a duration because
no "deactive" event was detected for them.  Fix this by assigning
duration of simulation length - start time for actions that do not have
a duration defined.  This will not effect actions that do not have a
duration (like Bolus, Consume Meal) because their respective build
functions default to duration = 0.