f/mmcdaniel-scenarioLoad: Remove redundant initialization
Remove code that loads two patients in a row at startup (this code was
needed when substance menu was not working correctly at startup).  Now
that menu items are only instantiated once, we do not need the
"initialized" role and can just use the onCompleted signal of each menu
item to determine whether it is active at startup and should create a
plot or not.