f/mmcdaniel-scenarioLoad: Scenarios now loaded from editor
One menu item: Launch Scenario Editor.  From the editor, users can add
actions/requests and may also load existing scenario files.

I think this means that I can get rid of the top level loader in the
window (introduced to make sure the the window had finished building
when loading scenarios the old way).  With this implementation, the
window is guaranteed to have been instantiated, so we should be able to
simplify things.