f/mmcdaniel-visuals: Improve scenario builder layout
Embed scenario builder content in a Page so that the tab bar can be
defined as Header content and the scenario buttons (Load, Reset, Save)
can be defined as Footer content.

Move scenario name and refresh rate options into header so that they are
available on both the scenario and data request tabs.  This gets rid of
the crowding effect caused by placing these items near the scenario
buttons.