f/mmcdaniel-visuals: Refactor action drawer
Refactoring action drawer so that only 1 dialog box is used, the
position of which is moved to be adjacent to the calling action.  The
dialog contains a loader that implements the appropriate biogears action
based on the selection from the menu.

Improve visual aspects of UISpinBox, UIComboBox, UIRadioButton, and
UITextField classes.  Changes also make these classes more broadly
applicable in other areas of the UI.

Make UISpinBox editable so that users can directly enter values instead
of clicking up/down arrows.